work on project page desktop
