--- src/App.css
@@ -1,8 +1,0 @@
-:root {
-  --main-yellow: #f8ed44;
-  --main-grey: #878680;
-  --main-black: black;
-  --white: white;
-  --main-blue: #a7e6e8;
-  --main-light-yellow: #fff0b3;
-}
