+++ src/assets/minus.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const Minus = () => (
+  <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 45.724 10">
+    <path id="Path_27" data-name="Path 27" d="M0,0H45.724" transform="translate(0 5)" fill="none" stroke="#6098BF" stroke-width="10"/>
+  </svg>
+)
+
+export default Minus;
