+++ src/assets/plus.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const Plus = () => (
+  <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 45.724 38.173">
+    <g id="Group_92" data-name="Group 92" transform="translate(-331.5 -998.327)">
+      <line id="Line_18" data-name="Line 18" y2="38.173" transform="translate(355.621 998.327)" fill="none" stroke="#6098BF" stroke-width="10"/>
+      <line id="Line_19" data-name="Line 19" x2="45.724" transform="translate(331.5 1017.413)" fill="none" stroke="#6098BF" stroke-width="10"/>
+    </g>
+  </svg>
+)
+
+export default Plus;
