--- src/assets/workEmailTemplate.js
@@ -1,9 +1,0 @@
-const template = ({ orgName, projectTitle }) => (`
-Hello ${orgName},\n\n
-My name is _____ and I am a ____. I am reaching out to you because I found your ${projectTitle} project listing on the Build for Black Lives project website. My teammates and I are very interested in working on your project, because _____.\n\n
-<Please remember to include information such as who your team is, what your background and experience is, and any questions you may have about the project, logistical or otherwise>\n\n
-Thank you very much and I hope to work with you soon!\n\n
-Best,
-`)
-
-export default template;
