+++ src/components/ProjectWork/workEmailTemplate.js
@@ -1,0 +1,10 @@
+const template = ({ orgName, projectTitle }) => (`
+Hello ${orgName},\n\n
+My name is _____ and I am a ____. I am reaching out to you because I found your ${projectTitle} project listing on the Build for Black Lives project website. My teammates (CC'd) and I would love to offer our ___[design/programming etc]____ skills to help work on your project pro-bono.\n\n
+<Please remember to include information such as who your team is, what your background and experience is, and any questions you may have about the project, logistical or otherwise>\n\n
+Please feel free to look at examples of my work at ___[portfolio link]___. You can contact me at ____[your email]___ or ___[your phone number]__. Thank you very much and I hope to work with you soon!\n\n
+Best,\n
+[Your name]
+`)
+
+export default template;
