+++ src/styling/project-work.css
@@ -1,29 +1,11 @@
-.project-work-page {
-  position: relative;
-}
-
-.confirmation-checks-modal-base {
-  position: fixed;
-  top: 72px;
-  left: 0;
-  height: calc(100vh - 72px);
-  width: 100vw;
-  background-color: rgba(1, 1, 1, 0.6);
-  align-content: center;
-}
-
-.confirmation-checks-modal {
-  color: black;
-  background-color: white;
-  border-radius: 10px;
-  padding: 24px;
-  width: 650px;
-  margin:  0 auto 16px auto;
+.collapsable-section {
+  color: var(--main-black);
+  margin-bottom: 24px;
 }
 
 .confirmation-checks-modal h2 {
   font-weight: bold;
-  font-size: 28px;
+  font-size: 24px;
   margin-bottom: 16px;
 }
 
@@ -59,6 +41,8 @@
 
 .confirmation-btn {
   margin-left: auto;
+  background-color: var(--main-blue);
+  font-weight: bold;
 }
 
 .confirmation-btn:disabled {
@@ -73,27 +57,52 @@
   margin-bottom: 48px;
 }
 
-.project-org-name {
+.project-work-page h3 {
   font-size: 24px;
+  margin-bottom: 4px;
+}
+
+.project-org-name {
+  font-size: 20px;
 }
 
 .project-detail-title {
   font-weight: bold;
-  color: var(--white);
 }
 
 .project-detail-content {
   font-size: 14px;
+  color: var(--main-black);
 }
 
 .project-detail-email {
-  background: var(--white);
-  padding: 24px;
-  color: black;
   white-space: pre-line;
   line-height: 18px;
 }
 
 .copy-button {
-  margin: 0 0 16px auto;
+  background-color: var(--main-blue);
+  font-weight: bold;
+  border: none;
+}
+
+.collapsable-section {
+  border-radius: 10px;
+}
+
+.collapsable-header {
+  display: flex;
+  cursor: pointer;
+}
+
+.collapsable-header-title {
+  margin-left: 16px;
+}
+
+.collapsable-content {
+  padding: 24px 60px;
+}
+
+.collapsable-content p {
+  color: var(--main-black);
 }
